// app/src/generated/prisma/models/SamlConfig.ts
export type SamlConfigModel = runtime.Types.Result.DefaultSelection<Prisma.$SamlConfigPayload>
// ...
export type $SamlConfigPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "SamlConfig"
  objects: {}
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    name: string
    idpEntityId: string
    idpSsoUrl: string
    idpSloUrl: string | null
    idpCertificate: string
    spEntityId: string
    emailAttribute: string
    displayNameAttr: string
    autoProvision: boolean
    active: boolean
    signRequests: boolean
    wantAssertionsSigned: boolean
    spPrivateKey: string | null
    spCertificate: string | null
    createdBy: string | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["samlConfig"]>
  composites: {}
}